+++ lib/ooxml_parser/pptx_parser/pptx_data/presentation/slide.rb
@@ -2,6 +2,7 @@ require_relative 'slide/presentation_alternate_content'
 require_relative 'slide/background'
 require_relative 'slide/common_slide_data'
 require_relative 'slide/connection_shape.rb'
+require_relative 'slide/presentation_notes.rb'
 require_relative 'slide/graphic_frame/graphic_frame'
 require_relative 'slide/slide/shapes_grouping'
 require_relative 'slide/slide/timing'
@@ -17,6 +18,8 @@ module OoxmlParser
     attr_reader :relationships
     # @return [String] name of slide
     attr_reader :name
+    # @return [Notes] note of slide
+    attr_reader :note
 
     def initialize(parent: nil, xml_path: nil)
       @parent = parent
@@ -59,7 +62,14 @@ module OoxmlParser
       end
       OOXMLDocumentObject.xmls_stack.pop
       @relationships = Relationships.parse_rels("#{OOXMLDocumentObject.path_to_folder}#{File.dirname(@xml_path)}/_rels/#{@name}.xml.rels")
+      parse_note
       self
+    end
+
+    def parse_note
+      notes_target = @relationships.target_by_type('notes')
+      return nil unless notes_target
+      @note = PresentationNotes.new(parent: self).parse("#{OOXMLDocumentObject.path_to_folder}#{File.dirname(@xml_path)}/#{notes_target}")
     end
   end
 end