+++ Changelog.md
@@ -27,6 +27,7 @@
 * Parsing `ChartAxis#scaling`
 * Parsing `ConnectionShape` in `ShapesGrouping` 
 * Parsing `Tab#leader`
+* Parsing Slide Notes
 
 ### Refactor
 * Change interface of `XlsxDrawing#graphic_frame`